backend-api/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.2.3

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 환경 변수 및 설정
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP 클라이언트
httpx==0.25.2
aiohttp==3.9.1

# AI API 클라이언트
google-generativeai==0.3.2
anthropic==0.7.8
openai==1.3.7

# 유틸리티
email-validator==2.1.0
python-slugify==8.0.1
pillow==10.1.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0

